+++ mempool/mempool_test.go
@@ -21,8 +21,7 @@ func newMempoolWithApp(cc proxy.ClientCreator) *Mempool {
 	appConnMem, _ := cc.NewABCIClient()
 	appConnMem.SetLogger(log.TestingLogger().With("module", "abci-client", "connection", "mempool"))
 	appConnMem.Start()
-	mempool := NewMempool(config.Mempool, appConnMem, 0)
-	mempool.SetLogger(log.TestingLogger())
+	mempool := NewMempool(config.Mempool, appConnMem, 0, log.TestingLogger())
 	return mempool
 }
 
@@ -104,7 +103,8 @@ func TestSerialReap(t *testing.T) {
 
 	mempool := newMempoolWithApp(cc)
 	appConnCon, _ := cc.NewABCIClient()
-	appConnCon.SetLogger(log.TestingLogger().With("module", "abci-client", "connection", "consensus"))
+	appConnCon.SetLogger(log.TestingLogger().With("module", "abci-client", "connection",
+		"consensus"))
 	if _, err := appConnCon.Start(); err != nil {
 		t.Fatalf("Error starting ABCI client: %v", err.Error())
 	}